Refactor exception package of postgresql-dialect-exception module
