+++ shardingsphere-dialect-exception/shardingsphere-postgresql-dialect-exception/src/main/java/org/apache/shardingsphere/dialect/postgresql/exception/authority/UnknownUsernameException.java
@@ -15,22 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.dialect.postgresql.exception;
+package org.apache.shardingsphere.dialect.postgresql.exception.authority;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.dialect.exception.SQLDialectException;
 
 /**
- * Privilege not granted exception.
+ * Unknown username exception.
  */
 @RequiredArgsConstructor
 @Getter
-public final class PrivilegeNotGrantedException extends SQLDialectException {
+public final class UnknownUsernameException extends SQLDialectException {
     
-    private static final long serialVersionUID = 8410672833723209253L;
+    private static final long serialVersionUID = -2895258747747089622L;
     
     private final String username;
-    
-    private final String databaseName;
 }
